move listing logic out to its own component cause I'm gonna use it again somewhere else. seriously why didn't I do this the first time I needed a newest feed
